Modified the fahrenheit to celcius converter program to use float

+++ 01/fahrenheit.c
@@ -5,7 +5,7 @@
  */
 
 int main() {
-  int fahr, celcius;
+  float fahr, celcius;
   int lower, upper, step;
 
   lower = 0;   /* lower limit of temperature table */
@@ -15,7 +15,7 @@ int main() {
   fahr = lower;
   while (fahr <= upper) {
     celcius = 5 * (fahr - 32) / 9;
-    printf("%d\t%d\n", fahr, celcius);
+    printf("%3.0f\t%6.1f\n", fahr, celcius);
     fahr += step;
   }
 }
